Dev editview

+++ TDIC/server-app/Controllers/api/ViewController.cs
@@ -26,7 +26,7 @@ namespace API.Controllers
 
         [AllowAnonymous]
         [HttpGet("details/id_article={id_article}&id_view={id_view}")]
-        public async Task<ActionResult> GetInstruction(long id_article,long id_view)
+        public async Task<ActionResult> GetInstruction(long id_article,int id_view)
         {
             return HandleResult(await Mediator.Send(new Details.Query{id_article = id_article,id_view=id_view}));
         }
